refactor: Split mock transaction structure to separate crate
This way if we only need MockTransaction data structure, we won't need
to include ckb-sdk which can pull in a lot of dependencies
