+++ ckb-sdk-types/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ckb-sdk"
+name = "ckb-sdk-types"
 version = "0.20.0"
 authors = ["Linfeng Qian <thewawar@gmail.com>", "Nervos Core Dev <dev@nervos.org>"]
 edition = "2018"
@@ -8,29 +8,11 @@ license = "MIT"
 [dependencies]
 serde = { version = "1.0", features = ["rc"] }
 serde_derive = "1.0"
-serde_json = "1.0"
-bech32 = "0.6.0"
-log = "0.4.6"
-jsonrpc-client-core = "0.5.0"
-jsonrpc-client-http = "0.5.0"
-secp256k1 = "0.15.0"
-faster-hex = "0.3"
-fnv = "1.0.3"
-aes-ctr = "0.3.0"
-scrypt = "0.2.0"
-rand = "0.6.5"
-tiny-keccak = "1.4"
-byteorder = "1.3.2"
-bitcoin_hashes = "0.3.2"
-uuid = { version = "0.7.4", features = ["v4"] }
-chrono = "0.4.6"
-failure = "0.1.5"
-lazy_static = "1.4.0"
 
 ckb-types = { git = "https://github.com/nervosnetwork/ckb", branch = "rc/v0.20" }
 ckb-script = { git = "https://github.com/nervosnetwork/ckb", branch = "rc/v0.20" }
 ckb-jsonrpc-types = { git = "https://github.com/nervosnetwork/ckb", branch = "rc/v0.20" }
 ckb-hash = { git = "https://github.com/nervosnetwork/ckb", branch = "rc/v0.20" }
-ckb-resource = { git = "https://github.com/nervosnetwork/ckb", branch = "rc/v0.20" }
+
+[dev-dependencies]
 ckb-crypto = { git = "https://github.com/nervosnetwork/ckb", branch = "rc/v0.20", features = ["secp"] }
-ckb-sdk-types = { path = "../ckb-sdk-types" }
